pyproject.toml
[project]
name = "etlproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Nicolas",email = "nicolascnagel@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "pyarrow (>=22.0.0,<23.0.0)",
    "azure-identity (>=1.25.1,<2.0.0)",
    "azure-storage-blob (>=12.27.1,<13.0.0)",
    "azure-keyvault (>=4.2.0,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pandera (>=0.27.0,<0.28.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "google-auth-oauthlib (>=1.2.3,<2.0.0)",
    "google-auth-httplib2 (>=0.2.1,<0.3.0)",
    "google-api-python-client (>=2.187.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
